+++ @geogebra/solver-sdk/src/paths/substitute-tree.ts
@@ -1,22 +1,20 @@
 import type { ExpressionTree } from '../parser/types';
 import { arrayToPath } from './array-to-path';
 
-/** Replace the subtree at the given path in `destination` with the given `substitution`
- * tree. By default, this will use the destination's decorators and omit the
- * substitution's decorators. Will automatically update the paths of the resulting tree.
+/** Replace the subtree at the given path in `destination` with the given
+ * `substitute` tree, using the substitute's decorators. Will automatically
+ * update the paths of the resulting tree.
  *
- * This method is meant to be used with filling in sub-steps of a solver solution into the
- * destination steps. For example, in "2*(3+4)", the inner step will have the fromExpr
- * "2+4". To show the entire fromExpr, we need to substitute it into the destination
- * expression at the right path. For fromExpr, we always need to keep the destination
- * decorators, but for toExpr it depends on the mathematical situation. Luckily, we can
- * just use the next step's fromExpr instead.
+ * This method is meant to be used with filling in sub-steps of a solver
+ * solution into the destination steps. For example, in "2*(3+4)", the inner
+ * step will have the fromExpr "(3+4)". To show the entire fromExpr, we need to
+ * substitute it into the destination expression at the right path. The solver
+ * includes any brackets that are needed when substituting the sub-steps.
  */
 export function substituteTree(
   destination: ExpressionTree,
-  substitution: ExpressionTree,
+  substitute: ExpressionTree,
   pathArray: number[],
-  useDestinationsDecorators = true,
 ): ExpressionTree {
   function inner(
     subDestination: ExpressionTree,
@@ -24,23 +22,15 @@ export function substituteTree(
     parentDestination: ExpressionTree | undefined = undefined,
   ): ExpressionTree {
     if (pathArray.length === pathIdx) {
-      const prefixedSubstitution = prefixPaths(substitution, arrayToPath(pathArray));
+      const prefixedSubstitute = prefixPaths(substitute, arrayToPath(pathArray));
       if (
-        !substitution.decorators?.length &&
-        parentDestination?.type === substitution.type &&
-        ['SmartProduct', 'Sum'].includes(substitution.type)
+        !substitute.decorators?.length &&
+        parentDestination?.type === substitute.type &&
+        ['SmartProduct', 'Sum'].includes(substitute.type)
       ) {
-        return { ...prefixedSubstitution, decorators: ['PartialBracket'] };
+        return { ...prefixedSubstitute, decorators: ['PartialBracket'] };
       }
-      if (useDestinationsDecorators) {
-        return {
-          ...prefixedSubstitution,
-          ...(subDestination.decorators?.length
-            ? { decorators: subDestination.decorators.slice() }
-            : null),
-        };
-      }
-      return prefixedSubstitution;
+      return prefixedSubstitute;
     }
     const index = pathArray[pathIdx];
     if (!('args' in subDestination) || index >= subDestination.args.length) {